reworking of user rights box [#157371124]

+++ app/assets/javascripts/catalog_manager/form.js.coffee
@@ -119,16 +119,19 @@ $ ->
 
   $(document).on 'click', '.remove-user-rights', (event) ->
     event.preventDefault()
-    identity_id = $(this).data('identity-id')
-    $.ajax
-      type: 'DELETE'
-      url: 'catalog_manager/user_right'
-      data:
-        user_rights:
-          identity_id: identity_id
-          organization_id: $(this).data('organization-id')
-      success: ->
-        $("#user-rights-row-#{identity_id}").fadeOut(1000, () -> $(this).remove())
+    if confirm (I18n['catalog_manager']['organization_form']['user_rights']['remove_confirm'])
+      identity_id = $(this).data('identity-id')
+      $.ajax
+        type: 'DELETE'
+        url: 'catalog_manager/user_right'
+        data:
+          user_rights:
+            identity_id: identity_id
+            organization_id: $(this).data('organization-id')
+
+  $(document).on 'click', '.cancel-user-rights', (event) ->
+    event.preventDefault()
+    $(this).closest('.row').fadeOut(1000, () -> $(this).remove())
 
   ##############################################
   ###          Service Components            ###
@@ -161,44 +164,7 @@ $ ->
   ##############################################
 
 
-
   ##############################################
   ###          Clinical Providers            ###
   ##############################################
-
-  # $(document).on 'change', '.super-user-checkbox', ->
-  #   $.ajax
-  #     type: if $(this).prop('checked') then 'POST' else 'DELETE'
-  #     url: '/catalog_manager/super_user'
-  #     data:
-  #       super_user:
-  #         identity_id: $(this).data('identity-id')
-  #         organization_id: $(this).data('organization-id')
-
-  $(document).on 'click', 'button.remove-clinical-provider', (event) ->
-    if confirm I18n["catalog_manager_js"]["clinical_provider_remove"]
-      $.ajax
-        type: 'DELETE'
-        url: '/catalog_manager/clinical_provider'
-        data:
-          clinical_provider:
-            identity_id: $(this).data('identity-id')
-            organization_id: $(this).data('organization-id')
-
-
-  # # clinical providers
-  # $('input#new_cp').live 'focus', -> $(this).val('')
-  # $('input#new_cp').live 'keydown.autocomplete', ->
-  #   $(this).autocomplete
-  #     source: "/catalog_manager/identities/search",
-  #     minLength: 3,
-  #     select: (event, ui) ->
-  #       $.post '/catalog_manager/identities/associate_with_org_unit', {identity: ui.item.value, org_unit: $('#org_unit_id').val(), rel_type: "clinical_provider_organizational_unit"}, (data) ->
-  #         $('#cp_info').html(data)
-
-  # $('.cp_delete').live 'click', ->
-  #   if confirm I18n["catalog_manager_js"]["clinical_provider_remove"]
-  #     $.post '/catalog_manager/identities/disassociate_with_org_unit', {relationship: $(this).attr('id'), org_unit: $('#org_unit_id').val(), rel_type: "clinical_provider_organizational_unit"}, (data) ->
-  #       $('#cp_info').html(data)
-
